revert []*Token -> State
For some mysterious reason, the github PR revert functionality
didn't work, so I'm manually reverting the change.  I experimented
a bit with a State struct that would contain both the tree
and the []*Token, but it didn't work out well.  There's not much
point to a custom type if I'm not going to use it.